backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
tiktoken
PyMuPDF
pypdf
chromadb
httpx
google-generativeai
